+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * nostr: add `Tags::dedup` method ([Yuki Kishimoto])
 * nostr: add `EncryptedSecretKey::decrypt` method ([Yuki Kishimoto])
 * nostr: add `Nip19Coordinate` struct ([Yuki Kishimoto])
+* nostr: add `Coordinate::verify` method ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])